2016/6/19 19:37 实现了商品的CRUD操作
在admin_manage.jsp中新增管理商品的界面。方便管理员管理商品，CRUD操作全在其中。
并将显示提示信息的函数又集成在一个函数中，在onLoad时就提示信息。

+++ src/main/java/service/ShoppingCarServiceImpl.java
@@ -39,16 +39,23 @@ public class ShoppingCarServiceImpl implements ShoppingCarService{
 	}
 
 	public boolean removeGoodFromCar(String userId, String goodId) {
-		
+		//如果用户没有买商品，就直接return
+		if(shoppingCarDao.getUserBoughtGood(userId, goodId) == null)
+			return false;
 		return shoppingCarDao.removeGoodFromCar(userId, goodId) > 0;
 	}
 
 	public boolean addGoodsBoughtNum(String userId, String goodId, int buyNum) {
+		if(shoppingCarDao.getUserBoughtGood(userId, goodId) == null) return false;
+		
 		return shoppingCarDao.addGoodsBoughtNum(userId, goodId, buyNum) > 0;
 	}
 
 	public boolean alterGoodsBoughtNum(String userId, String goodId, int buyNum) {
+		if(shoppingCarDao.getUserBoughtGood(userId, goodId) == null) return false;
 		return shoppingCarDao.alterGoodsBoughtNum(userId, goodId, buyNum) > 0;
 	}
+
+
 
 }
